amendments from rahul's comments. should be good to go now
